requirements.txt
semantic-kernel>=0.9.0
python-dotenv>=1.0.0
aiohttp>=3.9.0